change a little display
1.更改价格显示到小数点后两位
2.优化图书详情页面的总宽度
3.调整主页分类的切换效果

+++ src/com/mypack/dao/ItemsDAO.java
@@ -30,7 +30,7 @@ public class ItemsDAO {
 				item.setName(rs.getString("name"));
 				item.setCity(rs.getString("city"));
 				item.setCount(rs.getInt("count"));
-				item.setPrice(rs.getInt("price"));
+				item.setPrice(rs.getFloat("price"));
 				item.setPicture(rs.getString("picture"));
 				list.add(item);
